Move to_value to runtime + rename to TaskSignalVec
